+++ src/tags/closing-balance.ts
@@ -1,4 +1,4 @@
-import {compareArrays} from './../utils';
+import compareArrays from '../utils/compare-arrays';
 import {colonSymbolCode} from './../tokens';
 import openingBalanceTag, {BalanceInfoTag} from './opening-balance';
 import {State} from './../typings';