+++ src/typings.ts
@@ -27,9 +27,11 @@ export interface BalanceInfo {
 }
 
 export interface Transaction {
+    id: string;
     code: string;
     fundsCode: string;
     isCredit: boolean;
+    isExpense: boolean;
     currency: string;
     description: string;
     amount: number;